+++ go-fiber-postgres/internal/app/requests/role_request.go
@@ -8,16 +8,16 @@ import (
 
 type RoleCreateRequest struct {
 	Name          string `json:"name" validate:"required,min=2,max=50"`
-	PermissionIDs []uint `json:"permission_ids" validate:"omitempty,dive,exists=permissions.id"`
+	PermissionIDs []uint64 `json:"permission_ids" validate:"omitempty,dive,exists=permissions.id"`
 }
 
 type RoleUpdateRequest struct {
 	Name          *string `json:"name" validate:"omitempty,min=2,max=50"`
-	PermissionIDs []uint  `json:"permission_ids" validate:"omitempty,dive,exists=permissions.id"`
+	PermissionIDs []uint64 `json:"permission_ids" validate:"omitempty,dive,exists=permissions.id"`
 }
 
 type RoleSyncPermissionsRequest struct {
-	PermissionIDs []uint `json:"permission_ids" validate:"required,dive,exists=permissions.id"`
+	PermissionIDs []uint64 `json:"permission_ids" validate:"required,dive,exists=permissions.id"`
 }
 
 type RoleIndexRequest struct {
